Finished up original implementation
Realized that I hadn't finished the original tests, went back and found
a couple of issues.  Also, the other approach fell apart pretty quickly.
I still feel like there's a balance between determining the next roll
(or 2 rolls) in the sequence and showing information at the Frame level,
but that'll be for another day...